Added plotting to aid registration

+++ src/gp_compressor.h
@@ -18,26 +18,26 @@ public:
 protected:
     pointcloud::Ptr cloud; // the input pointcloud with RGB color information
     gp_octree octree;
-    float res;
+    double res;
     int sz;
 
     // the rotations of the patches
-    std::vector<Eigen::Quaternionf, Eigen::aligned_allocator<Eigen::Quaternionf> > rotations;
+    std::vector<Eigen::Quaterniond, Eigen::aligned_allocator<Eigen::Quaterniond> > rotations;
     // the 3D means of the patches
-    std::vector<Eigen::Vector3f, Eigen::aligned_allocator<Eigen::Vector3f> > means;
+    std::vector<Eigen::Vector3d, Eigen::aligned_allocator<Eigen::Vector3d> > means;
     // the means in RGB color space
-    //std::vector<Eigen::Vector3f, Eigen::aligned_allocator<Eigen::Vector3f> > RGB_means;
+    //std::vector<Eigen::Vector3d, Eigen::aligned_allocator<Eigen::Vector3d> > RGB_means;
 
-    std::vector<std::list<Eigen::Vector3f, Eigen::aligned_allocator<Eigen::Vector3f> > > S; // more or less for debugging
-    std::vector<std::list<Eigen::Vector3f, Eigen::aligned_allocator<Eigen::Vector3f> > > to_be_added;
+    std::vector<std::list<Eigen::Vector3d, Eigen::aligned_allocator<Eigen::Vector3d> > > S; // more or less for debugging
+    std::vector<std::list<Eigen::Vector3d, Eigen::aligned_allocator<Eigen::Vector3d> > > to_be_added;
     // same for RGB, first the R vectors then G and B
-    //Eigen::Matrix<float, Eigen::Dynamic, Eigen::Dynamic> RGB;
+    //Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic> RGB;
     // the masks showing where in the patches there are observations
     Eigen::Array<bool, Eigen::Dynamic, Eigen::Dynamic> W;
     std::vector<sparse_gp> gps;
 
-    void compute_rotation(Eigen::Matrix3f& R, const Eigen::MatrixXf& points);
-    void project_points(Eigen::Vector3f& center, const Eigen::Matrix3f& R, Eigen::MatrixXf& points,
+    void compute_rotation(Eigen::Matrix3d& R, const Eigen::MatrixXd& points);
+    void project_points(Eigen::Vector3d& center, const Eigen::Matrix3d& R, Eigen::MatrixXd& points,
                         const Eigen::Matrix<short, Eigen::Dynamic, Eigen::Dynamic>& colors,
                         const std::vector<int>& index_search, int* occupied_indices, int i);
     void project_cloud();
@@ -45,7 +45,7 @@ protected:
     void compress_colors();
     void train_processes();
 public:
-    gp_compressor(pointcloud::ConstPtr ncloud, float res = 0.1f, int sz = 10);
+    gp_compressor(pointcloud::ConstPtr ncloud, double res = 0.1f, int sz = 10);
     void save_compressed(const std::string& name);
     pointcloud::Ptr load_compressed();
 };
